Passing fp_conv_3_3 test due to coreir floating point reset fix

+++ apps/hardware_benchmarks/tests/fp_conv_3_3/process.cpp
@@ -11,7 +11,7 @@ using namespace Halide::Runtime;
 
 int main(int argc, char **argv) {
 
-  OneInOneOut_ProcessController<uint16_t> processor("fp_conv_3_3",
+  OneInOneOut_ProcessController<uint8_t> processor("fp_conv_3_3",
                                             {
                                               {"cpu",
                                                   [&]() { fp_conv_3_3(processor.input, processor.output); }
@@ -24,21 +24,9 @@ int main(int argc, char **argv) {
 
                                             });
 
-  processor.input = Buffer<uint16_t>(64, 64);
-  processor.output = Buffer<uint16_t>(62, 62);
+  processor.input = Buffer<uint8_t>(64, 64);
+  processor.output = Buffer<uint8_t>(62, 62);
   
   processor.process_command(argc, argv);
-
-
-  // Buffer<uint8_t> old_input;
-  // Buffer<uint8_t> new_input(64, 64);
-  // 
-  // old_input = load_and_convert_image("input.png");
-  // for (int i = 0; i < processor.input.width(); ++i) {
-  //   for (int j = 0; j < processor.input.height(); ++j) {
-  //     new_input(i, j) = i+j;
-  //   }
-  // }
-  // convert_and_save_image(new_input, "new_input.png");
   
 }
